revisão prova 1

+++ aula05/ex08.cpp
@@ -21,7 +21,7 @@ int main(){
             menor = vet[i];
     }
     
-    media = soma / VET;
+    media = (float)soma / VET;
 
     cout << "Somatoria " << soma << endl;
     cout << "Maior " << maior << endl;
